// src/Web/src/components/groups/components/invite/enums/CssClass.ts
export enum CssClass {
	Container = "addGroupContainerSecond",
	Inputs = "addGroupSecondSide__inputs",
	UserList = "addGroupSecondSide__userList",
	UserListItem = "addGroupSecondSide__userListItem",
	UserListItemName = "addGroupSecondSide__userListItemName",
	ButtonsGroup = "addGroupSecondSide__buttonsGroup",
}