// src/Web/src/components/groups/components/invite/enums/Resource.ts
export enum Resource {
	AddBtn = "addBtn",
	PrevBtn = "prevBtn",
	CreateBtn = "groups.addGroupForm.createBtn",
	EmailInput = "groups.addGroupForm.email",
	EmailInputComment = "groups.addGroupForm.emailComment",
	NameInput = "groups.addGroupForm.name",
	SurnameInput = "groups.addGroupForm.surname",
	BasicInfo = "groups.addGroupForm.basicInfo2",
	Confirm = "groups.addGroupForm.confirmBtn",
}